Move 4.x series document link to the page bottom.
I'd always thought that the expanded 'Getting Starting With The API' and 'Getting Started With Scripting' sections of the documentation page related to the Chaiscript 4.x series. This is due to indention of the panel divs making it look like they are children of the alert-success div related to 4.x documentation at the top of the page.

This change simply moves the 4.x documentation link to the bottom of the page to make it clearer.